Segmented layers but bad implementation.
