+++ app/src/main/java/com/example/mvvm_example/MainActivity.java
@@ -3,7 +3,6 @@ package com.example.mvvm_example;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.databinding.DataBindingUtil;
 import androidx.databinding.ObservableField;
-import androidx.lifecycle.ViewModelProvider;
 
 import android.os.Bundle;
 import android.view.View;
@@ -17,7 +16,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     //El binding correspondiente sera el camel case dela activity misma que tiene las suscripciones.
     ActivityMainBinding binding;
     //creamos un objeto GLOBAL ya que las suscripciones estaran ligadas con la direccion de memoria de este objeto en especifico.
-    User usuario = new User();
+    UserMVVM usuario = new UserMVVM();
     //creamos la variable String observable la cual mandaremos a travez del setter para modificar activamente la vista.
     ObservableField<String> name=new ObservableField<>();
 
@@ -35,10 +34,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         usuario.setName(name);
 
         //a nuestros variables (variables a los cuales estan suscritos nuestras vistas) les asiganamos el objeto con el que estos trabajaran (objetos a los cuales se suuscribiran)
-        binding.setUser(usuario);
-        //En el caso de el evento, le pasaremos la case contenedora de este evento (clase contenedora de eventos a la cual estara susrita)
-        binding.setHandler(new EventHandler(this));
-
+        binding.setUserMVVM(usuario);
         //este boton solo lo utilizaremos para comprobar que la vista esta tomando sus variables directo del objeto y no a partir de algun view.setText()
         btnCambiaVariable = findViewById(R.id.button);
         btnCambiaVariable.setOnClickListener(this);
@@ -46,8 +42,5 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     @Override
     public void onClick(View view) {
-        //modificamos el objeto inicial al cual nustra vista esta suscrito.
-        name.set("HI");
-        usuario.setName(name);
     }
 }
